--- src/styles/components/widgets/pop_up/label_popup/__init__.py
@@ -1,3 +1,0 @@
-from .object import (
-    LABEL_POP_UP_STYLESHEET,
-)
