--- src/styles/components/widgets/pop_up/label_popup/object.py
@@ -1,4 +1,0 @@
-from src.styles.stylesheet import stylesheet
-
-
-LABEL_POP_UP_STYLESHEET = stylesheet(__file__)
