--- src/styles/components/widgets/pop_up/label_popup/style.css
@@ -1,9 +1,0 @@
-QLabel{
-    border: none;
-    background-color: transparent;
-    font-size: 14px;
-    padding-left: 5px;
-    color: #28bfc9;
-    font-weight: 600;
-
-}
