+++ src/styles/components/windows/pop_up/label_popup/__init__.py
@@ -1,0 +1,3 @@
+from .object import (
+    DEFAULT_LABEL_POP_UP_STYLESHEET,
+)
