+++ src/styles/components/windows/pop_up/label_popup/object.py
@@ -1,0 +1,4 @@
+from src.styles.stylesheet import stylesheet
+
+
+DEFAULT_LABEL_POP_UP_STYLESHEET = stylesheet(__file__)
